correct data format in purkinje keyword Fixes #178

+++ ansys/heart/writer/custom_dynalib_keywords/keywords/em_ep_purkinje_network.py
@@ -22,7 +22,7 @@ class EmEpPurkinjeNetwork(KeywordBase):
                     Field("mid", int, 30, 10, kwargs.get("mid")),
                     Field("pointstx", float, 40, 10, kwargs.get("pointstx")),
                     Field("pointsty", float, 50, 10, kwargs.get("pointsty")),
-                    Field("pointstz", int, 60, 10, kwargs.get("pointstz")),
+                    Field("pointstz", float, 60, 10, kwargs.get("pointstz")),
                     Field("edgelen", float, 70, 10, kwargs.get("edgelen", 3)),
                 ],
             ),
@@ -92,12 +92,12 @@ class EmEpPurkinjeNetwork(KeywordBase):
         self._cards[0].set_value("pointsty", value)
 
     @property
-    def pointstz(self) -> typing.Optional[int]:
+    def pointstz(self) -> typing.Optional[float]:
         """Get or set the Z coordinate of the tree origin"""  # nopep8
         return self._cards[0].get_value("pointstz")
 
     @pointstz.setter
-    def pointstz(self, value: int) -> None:
+    def pointstz(self, value: float) -> None:
         self._cards[0].set_value("pointstz", value)
 
     @property
